Mori

+++ Cashfy/src/business/model/memento/FOMemento.java
@@ -6,13 +6,13 @@ import business.model.FinancialOption;
  * Implementa o padrão de projeto Memento
  */
 public class FOMemento {
-    private final int state; // O estado será o valor da opção financeira, pois ela varia com a tempo
+    private final double state; // O estado será o valor da opção financeira, pois ela varia com a tempo
 
-    public FOMemento(int state) {
+    public FOMemento(double state) {
         this.state = state;
     }
 
-    public int getState() {
+    public double getState() {
         return state;
     }
 }
